+++ src/RcppExports.cpp
@@ -91,6 +91,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// kursawe
+NumericVector kursawe(arma::vec x);
+RcppExport SEXP _smoof_kursawe(SEXP xSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::vec >::type x(xSEXP);
+    rcpp_result_gen = Rcpp::wrap(kursawe(x));
+    return rcpp_result_gen;
+END_RCPP
+}
 
 RcppExport SEXP evaluateBBOBFunctionCPP(SEXP, SEXP, SEXP, SEXP);
 RcppExport SEXP evaluateUFFunction(SEXP, SEXP, SEXP);
@@ -113,6 +124,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_smoof_dtlz_5", (DL_FUNC) &_smoof_dtlz_5, 2},
     {"_smoof_dtlz_6", (DL_FUNC) &_smoof_dtlz_6, 2},
     {"_smoof_dtlz_7", (DL_FUNC) &_smoof_dtlz_7, 2},
+    {"_smoof_kursawe", (DL_FUNC) &_smoof_kursawe, 1},
     {"evaluateBBOBFunctionCPP",      (DL_FUNC) &evaluateBBOBFunctionCPP,      4},
     {"evaluateUFFunction",           (DL_FUNC) &evaluateUFFunction,           3},
     {"getOptimumForBBOBFunctionCPP", (DL_FUNC) &getOptimumForBBOBFunctionCPP, 3},